Restructured component.transformation values.
Eventhough the UFO stores 6-tuple values for component transformations,
the RComponent object only offers scale and offset. This change stores
the transformation as 6-tuple, and makes the scale and offset values
available if requested This should not break scripts that rely on
component.offset or component.scale.
This allows components to store transformations with rotation or skew.