Fixes to UFO import scripts
Changed GetFolder to GetFileOrFolder, as this seems to cause the UI not
to grey out UFOs when it shouldn't be doing so. Only script that uses
GetFolder now is the script to import all UFOs from a folder (which
seems correct). Cleaned up duplicate Message imports. Added a PutFile
import to ExportFontToUFO, as it was missing this.